requirements.txt
openai>=1.40.0
langchain>=0.3.0
wikipedia>=1.4.0
requests>=2.31.0
gradio>=4.44.0
python-dotenv>=1.0.1
